fix: address PR comment about amount update inference
