+++ app/components/Views/confirmations/hooks/custom-amount/useCustomAmountStage.ts
@@ -1,11 +1,4 @@
-import {
-  Dispatch,
-  SetStateAction,
-  useCallback,
-  useEffect,
-  useRef,
-  useState,
-} from 'react';
+import { Dispatch, SetStateAction, useEffect, useRef, useState } from 'react';
 import {
   useIsTransactionPayQuoteLoading,
   useTransactionPayPrimaryRequiredToken,
@@ -25,9 +18,9 @@ export enum CustomAmountStage {
  * Owns the rendering state machine for `CustomAmountInfo`.
  *
  * Stage is computed from two layers: a stateful override (`AmountInput` when
- * the keyboard is open, `Loading` while an amount update is in flight) set by
- * the lifecycle controls or `setStage`; and a pure derivation from reactive
- * inputs (quotes, prefill flags) used whenever the override is `null`.
+ * the keyboard is open, `Loading` while an amount update is in flight) set via
+ * `setStage`; and a pure derivation from reactive inputs (quotes, prefill flags)
+ * used whenever the override is `null`.
  * The hook reads quote state itself so the component renders purely
  * off the returned `stage`.
  *
@@ -39,8 +32,8 @@ export enum CustomAmountStage {
  * @param options.isDepositPrefillEnabled - Whether deposit prefill is enabled.
  * @param options.isDepositPrefillLoading - Whether a deposit prefill is loading.
  * @param options.skipDepositPrefill - Whether deposit prefill is skipped.
- * @returns The current stage, amount-update lifecycle controls and state, and
- * a setter to override the stage.
+ * @returns The current stage, whether the amount is updating, and a setter to
+ * override the stage.
  */
 export function useCustomAmountStage({
   amountFiat,
@@ -61,8 +54,6 @@ export function useCustomAmountStage({
   isDepositPrefillLoading: boolean;
   skipDepositPrefill: boolean;
 }): {
-  beginAmountUpdate: () => boolean;
-  endAmountUpdate: () => void;
   isAmountUpdating: boolean;
   stage: CustomAmountStage;
   setStage: Dispatch<SetStateAction<CustomAmountStage | null>>;
@@ -74,24 +65,6 @@ export function useCustomAmountStage({
       ? CustomAmountStage.AmountInput
       : CustomAmountStage.Loading,
   );
-  const [isAmountUpdatePending, setIsAmountUpdatePending] = useState(false);
-  const isAmountUpdateInProgressRef = useRef(false);
-
-  const beginAmountUpdate = useCallback(() => {
-    if (isAmountUpdateInProgressRef.current) {
-      return false;
-    }
-
-    isAmountUpdateInProgressRef.current = true;
-    setIsAmountUpdatePending(true);
-    setStage(CustomAmountStage.Loading);
-    return true;
-  }, []);
-
-  const endAmountUpdate = useCallback(() => {
-    isAmountUpdateInProgressRef.current = false;
-    setIsAmountUpdatePending(false);
-  }, []);
 
   const isQuotesLoading = useIsTransactionPayQuoteLoading();
   const quotesLastUpdated = useTransactionPayQuotesLastUpdated();
@@ -106,7 +79,6 @@ export function useCustomAmountStage({
   // quote, not a stale one predating the amount update.
   const loadingBaselineRef = useRef<number | undefined>(undefined);
   const wasLoadingRef = useRef(false);
-  const settleOnCommitRef = useRef(false);
   // `amountFiat` from the previous commit, to recognise a no-op re-commit.
   const lastCommittedFiatRef = useRef<string | undefined>(undefined);
 
@@ -119,16 +91,14 @@ export function useCustomAmountStage({
    * amount. Compare the new amount against the *prior* commit ONLY here, before
    * overwriting the ref — a no-op re-commit fetches nothing, so clear at once.
    *
-   * Settle (later renders): after the amount commit resolves, clear on
-   * `hasFreshQuote` or `isQuotesLoading`. Key off *current* loading, not a
-   * latch: the controller pulses `isLoading` for an empty pre-fetch update
-   * before the real fetch, and a latch would clear in the gap and briefly
-   * derive `NoQuote`.
+   * Settle (later renders): clear on `hasFreshQuote` or `isQuotesLoading`. Key
+   * off *current* loading, not a latch: the controller pulses `isLoading` for
+   * an empty pre-fetch update before the real fetch, and a latch would clear in
+   * the gap and briefly derive `NoQuote`.
    */
   useEffect(() => {
     if (stageOverride !== CustomAmountStage.Loading) {
       wasLoadingRef.current = false;
-      settleOnCommitRef.current = false;
       return;
     }
 
@@ -138,21 +108,14 @@ export function useCustomAmountStage({
       wasLoadingRef.current = true;
       loadingBaselineRef.current = quotesLastUpdated;
       lastCommittedFiatRef.current = amountFiat;
-      settleOnCommitRef.current =
-        isNoOpRecommit || Boolean(disablePay) || hasPrefetchedQuote;
 
       // `disablePay` flows are direct transfers: no pay token, no quote, and the
-      // required-token amount may never resolve. Once the amount commit ends,
-      // there is nothing else to await. The same applies to no-op re-commits and
-      // amounts that already have a prefetched quote.
-      if (settleOnCommitRef.current && !isAmountUpdatePending) {
+      // required-token amount may never resolve. There is nothing to await once
+      // the amount is committed, so settle on the arm frame itself — the settle
+      // branch below is never re-entered when no reactive input changes.
+      if (isNoOpRecommit || disablePay || hasPrefetchedQuote) {
         setStage(null);
       }
-      return;
-    }
-
-    if (settleOnCommitRef.current && !isAmountUpdatePending) {
-      setStage(null);
       return;
     }
 
@@ -164,11 +127,7 @@ export function useCustomAmountStage({
       (loadingBaselineRef.current === undefined ||
         quotesLastUpdated > loadingBaselineRef.current);
 
-    if (
-      !isAmountUpdatePending &&
-      hasAmount &&
-      (hasFreshQuote || isQuotesLoading)
-    ) {
+    if (hasAmount && (hasFreshQuote || isQuotesLoading)) {
       setStage(null);
     }
   }, [
@@ -178,7 +137,6 @@ export function useCustomAmountStage({
     hasAmount,
     hasPrefetchedQuote,
     hasQuotes,
-    isAmountUpdatePending,
     isQuotesLoading,
     quotesLastUpdated,
   ]);
@@ -220,18 +178,8 @@ export function useCustomAmountStage({
     stage = CustomAmountStage.NoQuote;
   }
 
-  // Keep the amount update active until its commit settles, even if an older
-  // quote request was already loading. Afterwards, remain active only until
-  // quote fetching takes over.
   const isAmountUpdating =
-    stage === CustomAmountStage.Loading &&
-    (isAmountUpdatePending || !isQuotesLoading);
+    stage === CustomAmountStage.Loading && !isQuotesLoading;
 
-  return {
-    beginAmountUpdate,
-    endAmountUpdate,
-    isAmountUpdating,
-    setStage,
-    stage,
-  };
+  return { isAmountUpdating, setStage, stage };
 }
